@@ -19,43 +19,54 @@
     "sam-test-project"
   ],
   "scripts": {
-    "audit": "npm audit --omit dev && npm audit --workspaces --omit dev",
+    "audit": "npm audit --omit dev fix && npm audit --workspaces --omit dev fix",
     "lint": "eslint .",
     "lintfix": "eslint --cache --fix .",
-    "test:lint": "eslint .",
-    "test:packages": "tap --coverage-report=html --no-browser --no-check-coverage */tests/**/*.test.ts **/tests/**/*.test.ts ",
-    "test": "npm run test:lint && npm run test:packages",
+    "test:packages": "c8 tap --test-regex='/tests/.+\\.test\\.ts' --test-ignore='.aws-sam'",
+    "coverage": "c8 report --reporter=text-lcov > coverage/lcov.info",
+    "test": "npm run lint && npm run test:packages && npm run coverage",
     "postversion": "./scripts/sync-macro-version.js",
-    "prepare": "test ! -d '.git' || is-ci || husky install"
+    "prepare": "test ! -d '.git' || is-ci || husky install",
+    "coverage:html": "c8 report --reporter=html"
   },
   "tap": {
     "node-arg": [
       "--no-warnings",
       "--loader=ts-node/esm",
-      "--loader=esmock",
-      "--experimental-specifier-resolution=node"
+      "--experimental-specifier-resolution=node",
+      "--trace-deprecation"
     ]
   },
   "devDependencies": {
+    "@aws-sdk/client-cloudwatch": "^3.290.0",
+    "@swc-node/core": "^1.10.1",
+    "@swc-node/register": "^1.6.2",
+    "@swc/core": "^1.3.37",
+    "@swc/helpers": "^0.4.14",
     "@tsconfig/node-lts-strictest-esm": "^18.12.1",
-    "@types/prettier": "2.6.0",
-    "@types/serverless": "^3.12.9",
-    "@types/tap": "^15.0.7",
-    "@typescript-eslint/parser": "^5.0.0",
+    "@types/serverless": "^3.12.11",
+    "@types/tap": "^15.0.8",
+    "@typescript-eslint/eslint-plugin": "^5.54.0",
+    "@typescript-eslint/parser": "^5.54.0",
+    "ajv": "^8.12.0",
+    "c8": "^7.13.0",
     "cloudform": "^7.4.2",
+    "coverage-node": "^8.0.0",
     "esbuild": "^0.17.10",
-    "eslint": "^8.30.0",
-    "eslint-config-standard-with-typescript": "^24.0.0",
-    "eslint-plugin-import": "^2.23.0",
+    "eslint": "^8.35.0",
+    "eslint-config-standard-with-typescript": "^34.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.6.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-standard": "^5.0.0",
-    "esmock": "^2.1.0",
-    "husky": "^6.0.0",
-    "lint-staged": "^11.0.0",
-    "prettier": "2.0.5",
-    "tap": "^16.3.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.2",
+    "regenerator-runtime": "^0.13.11",
+    "serverless": "^3.28.1",
+    "tap": "^16.3.4",
     "ts-node": "^10.9.1",
-    "typescript": "~3.9.7"
+    "type-fest": "^3.6.1",
+    "typescript": "^4.9.5"
   },
   "engines": {
     "node": ">=12.0"
@@ -69,7 +80,6 @@
   },
   "dependencies": {
     "@types/lodash": "^4.14.191",
-    "@types/yaml": "^1.9.7",
     "case": "^1.6.3",
     "lodash": "^4.17.21",
     "yaml": "^1.10.2"
