@@ -2,29 +2,25 @@
 'use strict'
 
 import { test } from 'tap'
-import CloudFormationTemplate from 'slic-watch-core/cf-template'
-import esmock from 'esmock'
-import { createRequire } from 'module'
-const require = createRequire(import.meta.url)
-const cdkStack = require('./resources/cdk-ecs-cf.json')
-
-const cfMacroHandler = await esmock('../index', {})
+import { getResourcesByType } from 'slic-watch-core/cf-template'
+import { handler } from '../index'
+import cdkStack from './resources/cdk-ecs-cf.json' assert { type: 'json'}
 
 /**
  * Test the synthesized output from the CDK ECS Stack in `cdk-test-project`
  */
-test('ECS CDK stack', async (t) => {
+test('ECS CDK stack', (t) => {
   const event = {
     fragment: cdkStack
   }
-  const handlerResponse = await cfMacroHandler.handler(event)
+  const handlerResponse = handler(event)
   t.equal(handlerResponse.status, 'success')
-  const transformedTemplate = CloudFormationTemplate(handlerResponse.fragment)
+  const compiledTemplate = handlerResponse.fragment
 
   test('alarms are generated', (t) => {
-    const alarms = transformedTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+    const alarms = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
     t.equal(Object.keys(alarms).length, 6)
-    const alarmNames = Object.values(alarms).map(alarm => alarm.Properties?.AlarmName).sort()
+    const alarmNames = Object.values(alarms).map(alarm => alarm.Properties?.AlarmName).sort((a: string, b: string) => a.localeCompare(b))
     t.same(alarmNames, [
       'ECS_CPUAlarm_${MyWebServerService2FE7341D.Name}',
       'ECS_MemoryAlarm_${MyWebServerService2FE7341D.Name}',
