@@ -2,142 +2,99 @@
 
 import { test } from 'tap'
 import _ from 'lodash'
-import esmock from 'esmock'
-import { CloudFormationTemplate } from 'slic-watch-core/cf-template'
-import { createRequire } from 'module'
-const require = createRequire(import.meta.url)
-const template = require('./event.json')
+import { handler } from '../index'
+import template from './event.json' assert { type: 'json'}
 
-const event = { fragment: template, templateParameterValues: { stack: 'sam-test-stack-project' } }
+const event = { fragment: template }
 
-type TestState = {
-  dashboardCalled?:boolean,
-  alarmsCalled?: boolean,
-  addDashboardCfTemplate?: CloudFormationTemplate,
-  addAlarmsCfTemplate?: CloudFormationTemplate
-}
+test('macro returns success', t => {
+  const result = handler(event)
+  t.equal(result.status, 'success')
+  t.end()
+})
 
-let testState: TestState
+test('macro uses SNS Topic environment variable if specified', t => {
+  process.env.ALARM_SNS_TOPIC = 'arn:aws:sns:eu-west-1:123456789123:TestTopic'
+  try {
+    const result = handler(event)
+    t.equal(result.status, 'success')
+  } finally {
+    delete process.env.ALARM_SNS_TOPIC
+  }
+  t.end()
+})
 
-const lambda = await esmock('../index', {
-  'slic-watch-core/dashboards/dashboard': () => {
-    testState.dashboardCalled = true
-    return {
-      addDashboard: (cfTemplate: CloudFormationTemplate) => {
-        testState.addDashboardCfTemplate = cfTemplate
-      }
-    }
-  },
-  'slic-watch-core/alarms/alarms': () => {
-    testState.alarmsCalled = true
-    return {
-      addAlarms: (cfTemplate) => {
-        testState.addAlarmsCfTemplate = cfTemplate
+test('macro uses topicArn if specified', t => {
+  const eventWithTopic = {
+    ...event,
+    fragment: {
+      ...event.fragment,
+      Metadata: {
+        ...event.fragment.Metadata,
+        slicWatch: {
+          ...event.fragment.Metadata.slicWatch,
+          topicArn: 'arn:aws:sns:eu-west-1:123456789123:TestTopic'
+        }
       }
     }
   }
+  const result = handler(eventWithTopic)
+  t.equal(result.status, 'success')
+  t.end()
 })
 
-test('index', t => {
-  t.beforeEach(t => {
-    t.ok(true)
-    testState = {}
-  })
+test('Macro skips SLIC Watch if top-level enabled==false', t => {
+  const testevent = _.cloneDeep(event)
+  testevent.fragment.Metadata.slicWatch.enabled = false
+  handler(testevent)
+  t.end()
+})
 
-  t.test('macro returns success', async t => {
-    const result = await lambda.handler(event, null)
-    t.equal(result.status, 'success')
-  })
-
-  t.test('macro uses SNS Topic environment variable if specified', async t => {
-    process.env.ALARM_SNS_TOPIC = 'arn:aws:sns:eu-west-1:123456789123:TestTopic'
-    try {
-      const result = await lambda.handler(event, null)
-      t.equal(result.status, 'success')
-    } finally {
-      delete process.env.ALARM_SNS_TOPIC
-    }
-  })
-
-  t.test('macro uses topicArn if specified', async t => {
-    const eventWithTopic = {
-      ...event,
-      fragment: {
-        ...event.fragment,
-        Metadata: {
-          ...event.fragment.Metadata,
-          slicWatch: {
-            ...event.fragment.Metadata.slicWatch,
-            topicArn: 'arn:aws:sns:eu-west-1:123456789123:TestTopic'
-          }
+test('Macro adds dashboard and alarms if no function configuration is provided', t => {
+  const testEvent = {
+    ...event,
+    fragment: {
+      ...event.fragment,
+      Resources: {
+        ...event.fragment.Resources,
+        HelloLambdaFunction: {
+          ...event.fragment.Resources.HelloLambdaFunction,
+          Metadata: {}
         }
       }
     }
-    const result = await lambda.handler(eventWithTopic, null)
-    t.equal(result.status, 'success')
-  })
+  }
+  const compiledTemplate = handler(testEvent).fragment
+  t.same(compiledTemplate.Resources.Properties, template.Resources.Properties)
+  t.end()
+})
 
-  t.test('Macro skips SLIC Watch if top-level enabled==false', async t => {
-    const testevent = _.cloneDeep(event)
-    testevent.fragment.Metadata.slicWatch.enabled = false
-    await lambda.handler(testevent, null)
-    t.notOk(testState.alarmsCalled)
-  })
+test('Macro execution fails if an invalid SLIC Watch config is provided', t => {
+  const testevent = _.cloneDeep(event)
+  testevent.fragment.Metadata.slicWatch.topicArrrrn = 'pirateTopic'
+  const result = handler(testevent)
+  t.equal(result.status, 'fail')
+  t.end()
+})
 
-  t.test('Macro adds dashboard and alarms', async t => {
-    await lambda.handler(event, null)
-    t.equal(testState.addDashboardCfTemplate?.getSourceObject(), template)
-    t.equal(testState.addAlarmsCfTemplate?.getSourceObject(), template)
-  })
+test('Macro execution succeeds with no slicWatch config', t => {
+  const testevent = _.cloneDeep(event)
+  delete testevent.fragment.Metadata.slicWatch
+  handler(testevent)
+  t.end()
+})
 
-  t.test('Macro adds dashboard and alarms if no function configuration is provided', async t => {
-    const testEvent = {
-      ...event,
-      fragment: {
-        ...event.fragment,
-        Resources: {
-          ...event.fragment.Resources,
-          HelloLambdaFunction: {
-            ...event.fragment.Resources.HelloLambdaFunction,
-            Metadata: {}
-          }
-        }
-      }
-    }
-    await lambda.handler(testEvent, null)
-    // @ts-ignore
-    t.equal(testState.addDashboardCfTemplate.getSourceObject().Resources.Properties, template.Resources.Properties)
-    // @ts-ignore
-    t.equal(testState.addAlarmsCfTemplate.getSourceObject().Resources.Properties, template.Resources.Properties)
-  })
+test('Macro execution succeeds if no SNS Topic is provided', t => {
+  const testevent = _.cloneDeep(event)
+  delete testevent.fragment.Metadata.slicWatch.topicArn
+  handler(testevent)
+  t.end()
+})
 
-  t.test('Macro execution fails if an invalid SLIC Watch config is provided', async t => {
-    const testevent = _.cloneDeep(event)
-    testevent.fragment.Metadata.slicWatch.topicArrrrn = 'pirateTopic'
-    const result = await lambda.handler(testevent, null)
-    t.equal(result.status, 'fail')
-  })
-
-  t.test('Macro execution succeeds with no slicWatch config', async t => {
-    const testevent = _.cloneDeep(event)
-    delete testevent.fragment.Metadata.slicWatch
-    await lambda.handler(testevent, null)
-    t.ok(testState.alarmsCalled)
-  })
-
-  t.test('Macro execution succeeds if no SNS Topic is provided', async t => {
-    const testevent = _.cloneDeep(event)
-    delete testevent.fragment.Metadata.slicWatch.topicArn
-    await lambda.handler(testevent, null)
-    t.ok(testState.alarmsCalled)
-  })
-
-  t.test('Macro succeeds with no custom section', async t => {
-    const testevent = _.cloneDeep(event)
-    delete testevent.fragment.Metadata
-    const result = await lambda.handler(testevent, null)
-    t.equal(result.status, 'success')
-  })
-
+test('Macro succeeds with no custom section', t => {
+  const testevent = _.cloneDeep(event)
+  delete testevent.fragment.Metadata
+  const result = handler(testevent)
+  t.equal(result.status, 'success')
   t.end()
 })
