@@ -3,33 +3,33 @@
 import _ from 'lodash'
 import Ajv from 'ajv'
 
-import alarms from 'slic-watch-core/alarms/alarms'
-import dashboard from 'slic-watch-core/dashboards/dashboard'
-import CloudFormationTemplate from 'slic-watch-core/cf-template'
+import addAlarms from 'slic-watch-core/alarms/alarms'
+import addDashboard from 'slic-watch-core/dashboards/dashboard'
 import defaultConfig from 'slic-watch-core/inputs/default-config'
 import { slicWatchSchema } from 'slic-watch-core/inputs/config-schema'
+import { getResourcesByType } from 'slic-watch-core/cf-template'
 import pino from 'pino'
 
 const logger = pino({ name: 'macroHandler' })
 
-type Event = {
-  requestId
-  status: string
+interface Event {
+  requestId?
+  status?: string
   fragment
 }
 
-type SlicWatchConfig = {
-  topicArn: string
-  ActionsEnabled: boolean
+interface SlicWatchConfig {
+  topicArn?: string
+  enabled?: boolean
 }
-export async function handler (event: Event) {
+export function handler (event: Event): Event {
   let status = 'success'
   logger.info({ event })
-  let outputFragment = event.fragment
+  const outputFragment = event.fragment
   try {
-    const slicWatchConfig:SlicWatchConfig = (outputFragment.Metadata || {}).slicWatch || {}
+    const slicWatchConfig: SlicWatchConfig = outputFragment.Metadata?.slicWatch ?? {}
 
-    if (slicWatchConfig.ActionsEnabled !== false) {
+    if (slicWatchConfig.enabled ?? true) {
       const ajv = new Ajv({
         unicodeRegExp: false
       })
@@ -42,45 +42,35 @@ export async function handler (event: Event) {
         throw new Error('SLIC Watch configuration is invalid: ' + ajv.errorsText(slicWatchValidate.errors))
       }
 
-      const alarmActions = []
-      if (slicWatchConfig.topicArn) {
-        // @ts-ignore
-        alarmActions.push(slicWatchConfig.topicArn)
-      } else if (process.env.ALARM_SNS_TOPIC) {
-        // @ts-ignore
-        alarmActions.push(process.env.ALARM_SNS_TOPIC)
-      }
+      const alarmActions: string[] = []
+      slicWatchConfig?.topicArn != null && alarmActions.push(slicWatchConfig.topicArn)
+      process.env.ALARM_SNS_TOPIC != null && alarmActions.push(process.env.ALARM_SNS_TOPIC)
 
       const context = {
         alarmActions
       }
-      const cfTemplate = CloudFormationTemplate(
-        outputFragment
-      )
       const functionAlarmConfigs = {}
       const functionDashboardConfigs = {}
 
-      const lambdaResources = cfTemplate.getResourcesByType(
-        'AWS::Lambda::Function'
+      const lambdaResources = getResourcesByType(
+        'AWS::Lambda::Function',
+        outputFragment
       )
 
       for (const [funcResourceName, funcResource] of Object.entries(lambdaResources)) {
-        const funcConfig = funcResource.Metadata?.slicWatch || {}
-        functionAlarmConfigs[funcResourceName] = funcConfig.alarms || {}
-        functionDashboardConfigs[funcResourceName] = funcConfig.dashboard || {}
+        const funcConfig = funcResource.Metadata?.slicWatch ?? {}
+        functionAlarmConfigs[funcResourceName] = funcConfig.alarms ?? {}
+        functionDashboardConfigs[funcResourceName] = funcConfig.dashboard ?? {}
       }
 
-      const alarmService = alarms(config.alarms, functionAlarmConfigs, context)
-      alarmService.addAlarms(cfTemplate)
-      const dashboardService = dashboard(config.dashboard, functionDashboardConfigs)
-      dashboardService.addDashboard(cfTemplate)
-      outputFragment = cfTemplate.getSourceObject()
+      addAlarms(config.alarms, functionAlarmConfigs, context, outputFragment)
+      addDashboard(config.dashboard, functionDashboardConfigs, outputFragment)
     }
   } catch (err) {
     console.error(err)
     status = 'fail'
   }
-  logger.info({ outputFragment })
+  // logger.info({ outputFragment })
 
   return {
     requestId: event.requestId,
