@@ -1,112 +1,60 @@
 'use strict'
 
-import { CloudFormationTemplate } from '../cf-template'
-import Resource from 'cloudform-types/types/resource'
-import { Context, createAlarm } from './default-config-alarms'
+import { getResourcesByType, addResource, type ResourceType } from '../cf-template'
+import { type Context, createAlarm, type DefaultAlarmsProperties } from './default-config-alarms'
 import { getStatisticName } from './get-statistic-name'
 import { makeResourceName } from './make-name'
-import { AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import { type AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import type Template from 'cloudform-types/types/template'
 
-export type AppSyncAlarmProperties = AlarmProperties & {
-  '5XXError': AlarmProperties
-  Latency: AlarmProperties
+export interface AppSyncAlarmProperties {
+  enabled?: boolean
+  '5XXError': DefaultAlarmsProperties
+  Latency: DefaultAlarmsProperties
 }
 
-export type AppSyncAlarm= AlarmProperties & {
+export type AppSyncAlarm = AlarmProperties & {
   AppSyncResourceName: string
 }
+
+type AppSyncMetrics = '5XXError' | 'Latency'
+
+const executionMetrics: AppSyncMetrics[] = ['5XXError', 'Latency']
 
 /**
  * appSyncAlarmProperties The fully resolved alarm configuration
  */
-export default function appSyncAlarms (appSyncAlarmProperties: AppSyncAlarmProperties, context: Context) {
-  return {
-    createAppSyncAlarms
-  }
-
+export default function createAppSyncAlarms (appSyncAlarmProperties: AppSyncAlarmProperties, context: Context, compiledTemplate: Template, additionalResources: ResourceType = {}) {
   /**
    * Add all required AppSync alarms to the provided CloudFormation template
    * based on the AppSync resources found within
    *
    * A CloudFormation template object
    */
-  function createAppSyncAlarms (cfTemplate: CloudFormationTemplate) {
-    const appSyncResources = cfTemplate.getResourcesByType(
-      'AWS::AppSync::GraphQLApi'
-    )
+  const appSyncResources = getResourcesByType('AWS::AppSync::GraphQLApi', compiledTemplate, additionalResources)
 
-    for (const [appSyncResourceName, appSyncResource] of Object.entries(appSyncResources)) {
-      const alarms = []
-      if (appSyncAlarmProperties['5XXError'].ActionsEnabled) {
-        alarms.push(create5XXAlarm(
-          appSyncResourceName,
-          appSyncResource,
-          appSyncAlarmProperties['5XXError']
-        ))
+  for (const [appSyncResourceName, appSyncResource] of Object.entries(appSyncResources)) {
+    for (const metric of executionMetrics) {
+      const config = appSyncAlarmProperties[metric]
+      if (config.enabled !== false) {
+        const graphQLName: string = appSyncResource.Properties?.Name
+        const threshold = config.Threshold
+        delete config.enabled
+        const appSyncAlarmProperties: AppSyncAlarm = {
+          AlarmName: `AppSync${metric}Alarm_${graphQLName}`,
+          AlarmDescription: `AppSync ${metric} ${getStatisticName(config)} for ${graphQLName} breaches ${threshold}`,
+          AppSyncResourceName: appSyncResourceName,
+          MetricName: metric,
+          Namespace: 'AWS/AppSync',
+          Dimensions: [
+            { Name: 'GraphQLAPIId', Value: `\${${appSyncResourceName}.ApiId}}` }
+          ],
+          ...config
+        }
+        const resourceName = makeResourceName('AppSync', graphQLName, metric)
+        const resource = createAlarm(appSyncAlarmProperties, context)
+        addResource(resourceName, resource, compiledTemplate)
       }
-
-      if (appSyncAlarmProperties.Latency.ActionsEnabled) {
-        alarms.push(createLatencyAlarm(
-          appSyncResourceName,
-          appSyncResource,
-          appSyncAlarmProperties.Latency
-        ))
-      }
-      for (const alarm of alarms) {
-        cfTemplate.addResource(alarm.resourceName, alarm.resource)
-      }
-    }
-  }
-
-  function create5XXAlarm (appSyncResourceName: string, appSyncResource: Resource, config: AlarmProperties) {
-    const graphQLName = appSyncResource.Properties.Name
-    const threshold = config.Threshold
-    const appSyncAlarmProperties:AppSyncAlarm = {
-      AlarmName: `AppSync5XXErrorAlarm_${graphQLName}`,
-      AlarmDescription: `AppSync 5XX Error ${getStatisticName(config)} for ${graphQLName} breaches ${threshold}`,
-      AppSyncResourceName: appSyncResourceName,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: '5XXError',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/AppSync',
-      Dimensions: [
-        { Name: 'GraphQLAPIId', Value: `\${${appSyncResourceName}.ApiId}}` }
-      ]
-    }
-    return {
-      resourceName: makeResourceName('AppSync', graphQLName, '5XXError'),
-      resource: createAlarm(appSyncAlarmProperties, context)
-    }
-  }
-
-  function createLatencyAlarm (appSyncResourceName: string, appSyncResource: Resource, config: AlarmProperties) {
-    const graphQLName = appSyncResource.Properties.Name
-    const threshold = config.Threshold
-    const appSyncAlarmProperties:AppSyncAlarm = {
-      AlarmName: `AppSyncLatencyAlarm_${graphQLName}`,
-      AlarmDescription: `AppSync Latency ${getStatisticName(config)} for ${graphQLName} breaches ${threshold}`,
-      AppSyncResourceName: appSyncResourceName,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: 'Latency',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/AppSync',
-      Dimensions: [
-        { Name: 'GraphQLAPIId', Value: `\${${appSyncResourceName}.ApiId}}` }
-      ]
-    }
-    return {
-      resourceName: makeResourceName('AppSync', graphQLName, 'Latency'),
-      resource: createAlarm(appSyncAlarmProperties, context)
     }
   }
 }
