@@ -1,80 +1,88 @@
-import { AlbTargetAlarmProperties, AlbTargetAlarm } from './alb-target-group'
-import { AlbAlarmProperties, AlbAlarm } from './alb'
-import { ApiAlarm, ApiGwAlarmProperties } from './api-gateway'
-import { AppSyncAlarm, AppSyncAlarmProperties } from './appsync'
-import { DynamoDbAlarmProperties } from './dynamodb'
-import { EcsAlarm, EcsAlarmsConfig } from './ecs'
-import { EventbridgeAlarm, EventsAlarmsConfig } from './eventbridge'
-import { KinesisAlarmProperties } from './kinesis'
-import { LambdaAlarm, LambdaFunctionAlarmPropertiess } from './lambda'
-import { SnsAlarm, SnsAlarmsConfig } from './sns'
-import { SqsAlarm, SqsAlarmsConfig } from './sqs'
-import { SfAlarmsConfig, SmAlarm } from './step-functions'
-import { AlarmsCascade } from '../inputs/cascading-config'
-import { AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+'use strict'
 
-export type ReturnAlarm = {
+import { type AlbTargetAlarmProperties, type AlbTargetAlarm } from './alb-target-group'
+import { type AlbAlarmProperties, type AlbAlarm } from './alb'
+import { type ApiAlarm, type ApiGwAlarmProperties } from './api-gateway'
+import { type AppSyncAlarm, type AppSyncAlarmProperties } from './appsync'
+import { type DynamoDbAlarmProperties } from './dynamodb'
+import { type EcsAlarm, type EcsAlarmsConfig } from './ecs'
+import { type EventbridgeAlarm, type EventsAlarmsConfig } from './eventbridge'
+import { type KinesisAlarmProperties } from './kinesis'
+import { type LambdaAlarm, type LambdaFunctionAlarmProperties } from './lambda'
+import { type SnsAlarm, type SnsAlarmsConfig } from './sns'
+import { type SqsAlarm, type SqsAlarmsConfig } from './sqs'
+import { type SfAlarmsConfig, type SmAlarm } from './step-functions'
+import { type SlicWatchAlarmsConfig } from '../inputs/cascading-config'
+import { type AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import type Resource from 'cloudform-types/types/resource'
+
+export interface ReturnResource {
   Type: string
   Properties: AlarmProperties
 }
 
-export type Context = {
+export interface ReturnAlarm {
+  resourceName: string
+  resource: Resource
+}
+
+// export interface SlicWatchAlarmProperties {
+//   enabled: boolean
+//   Period: number
+//   EvaluationPeriods: number
+//   TreatMissingData: string
+//   ComparisonOperator: string
+// }
+
+export interface DefaultAlarmsProperties extends AlarmProperties {
+  enabled?: boolean
+}
+
+export interface Context {
   alarmActions: string[]
 }
 
-type AllAlarms = AlarmProperties | AlbAlarm | AlbTargetAlarm | ApiAlarm | AppSyncAlarm | EcsAlarm | EventbridgeAlarm | LambdaAlarm | SnsAlarm | SqsAlarm | SmAlarm
+export type AllAlarms = AlarmProperties | AlbAlarm | AlbTargetAlarm | ApiAlarm | AppSyncAlarm | EcsAlarm | EventbridgeAlarm | LambdaAlarm | SnsAlarm | SqsAlarm | SmAlarm
 
-export function createAlarm(alarm: AlarmProperties, context?: Context): ReturnAlarm
-export function createAlarm(alarm: AlbAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: AlbTargetAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: ApiAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: AppSyncAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: EcsAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: EventbridgeAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: LambdaAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: SnsAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: SqsAlarm, context?: Context): ReturnAlarm
-export function createAlarm(alarm: SmAlarm, context?: Context): ReturnAlarm
-export function createAlarm (alarm:AllAlarms, context?: Context): ReturnAlarm {
+export function createAlarm (alarm: AlarmProperties, context?: Context): ReturnResource
+export function createAlarm (alarm: AlbAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: AlbTargetAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: ApiAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: AppSyncAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: EcsAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: EventbridgeAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: LambdaAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: SnsAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: SqsAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: SmAlarm, context?: Context): ReturnResource
+export function createAlarm (alarm: AllAlarms, context?: Context): ReturnResource {
   return {
     Type: 'AWS::CloudWatch::Alarm',
     Properties: {
       ActionsEnabled: true,
-      AlarmActions: context.alarmActions,
-      AlarmName: alarm.AlarmName,
-      AlarmDescription: alarm.AlarmDescription,
-      EvaluationPeriods: alarm.EvaluationPeriods,
-      ComparisonOperator: alarm.ComparisonOperator,
-      Threshold: alarm.Threshold,
-      TreatMissingData: alarm.TreatMissingData,
-      Dimensions: alarm.Dimensions,
-      Metrics: alarm.Metrics,
-      MetricName: alarm.MetricName,
-      Namespace: alarm.Namespace,
-      Period: alarm.Period,
-      Statistic: alarm.Statistic,
-      ExtendedStatistic: alarm.ExtendedStatistic
+      AlarmActions: context?.alarmActions,
+      ...alarm
     }
   }
 }
 
-export type AllAlarmsConfig = {
-  ActionsEnabled: boolean
-  alarms?: AlarmsCascade
+export interface AllAlarmsConfig {
+  enabled: boolean
+  alarms?: SlicWatchAlarmsConfig
 }
 
 export type AlarmsConfig = AlbTargetAlarmProperties & AlbAlarmProperties & ApiGwAlarmProperties & AppSyncAlarmProperties & DynamoDbAlarmProperties
-& EcsAlarmsConfig & EventsAlarmsConfig & KinesisAlarmProperties & LambdaFunctionAlarmPropertiess & SnsAlarmsConfig & SqsAlarmsConfig & SfAlarmsConfig
+& EcsAlarmsConfig & EventsAlarmsConfig & KinesisAlarmProperties & LambdaFunctionAlarmProperties & SnsAlarmsConfig & SqsAlarmsConfig & SfAlarmsConfig
 
-export type FunctionAlarmPropertiess = {
-  HelloLambdaFunction?: LambdaFunctionAlarmPropertiess
-  ThrottlerLambdaFunction?: LambdaFunctionAlarmPropertiess
-  DriveStreamLambdaFunction?: LambdaFunctionAlarmPropertiess
-  DriveQueueLambdaFunction?: LambdaFunctionAlarmPropertiess
-  DriveTableLambdaFunction?: LambdaFunctionAlarmPropertiess
-  StreamProcessorLambdaFunction?: LambdaFunctionAlarmPropertiess
-  HttpGetterLambdaFunction?: LambdaFunctionAlarmPropertiess
-  SubscriptionHandlerLambdaFunction?: LambdaFunctionAlarmPropertiess
-  EventsRuleLambdaFunction?: LambdaFunctionAlarmPropertiess
-  AlbEventLambdaFunction?: LambdaFunctionAlarmPropertiess
+export interface FunctionAlarmProperties {
+  HelloLambdaFunction?: LambdaFunctionAlarmProperties
+  ThrottlerLambdaFunction?: LambdaFunctionAlarmProperties
+  DriveStreamLambdaFunction?: LambdaFunctionAlarmProperties
+  DriveQueueLambdaFunction?: LambdaFunctionAlarmProperties
+  DriveTableLambdaFunction?: LambdaFunctionAlarmProperties
+  StreamProcessorLambdaFunction?: LambdaFunctionAlarmProperties
+  HttpGetterLambdaFunction?: LambdaFunctionAlarmProperties
+  SubscriptionHandlerLambdaFunction?: LambdaFunctionAlarmProperties
+  EventsRuleLambdaFunction?: LambdaFunctionAlarmProperties
+  AlbEventLambdaFunction?: LambdaFunctionAlarmProperties
 }
