@@ -1,19 +1,22 @@
 'use strict'
 
-import { CloudFormationTemplate } from '../cf-template'
-import Resource from 'cloudform-types/types/resource'
-import { Context, createAlarm } from './default-config-alarms'
-import { AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import { getResourcesByType, addResource, type ResourceType } from '../cf-template'
+import { type Context, createAlarm, type DefaultAlarmsProperties } from './default-config-alarms'
+import { type AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import type Template from 'cloudform-types/types/template'
 
-export type EcsAlarmsConfig = AlarmProperties & {
-  MemoryUtilization: AlarmProperties
-  CPUUtilization: AlarmProperties
+export interface EcsAlarmsConfig {
+  enabled?: boolean
+  MemoryUtilization: DefaultAlarmsProperties
+  CPUUtilization: DefaultAlarmsProperties
 }
 
 export type EcsAlarm = AlarmProperties & {
-  ServiceName: string,
+  ServiceName: string
   ClusterName: string
 }
+
+type EcsMEtrics = 'MemoryUtilization' | 'CPUUtilization'
 
 /**
  * Given CloudFormation syntax for an ECS cluster, derive CloudFormation syntax for
@@ -22,115 +25,58 @@ export type EcsAlarm = AlarmProperties & {
  * @param  cluster CloudFormation syntax for an ECS cluster
  * @returns CloudFormation syntax for the cluster's name
  */
-export function resolveEcsClusterNameAsCfn (cluster) {
+export function resolveEcsClusterNameAsCfn (cluster): any {
   if (typeof cluster === 'string') {
     if (cluster.startsWith('arn:')) {
-      return cluster.split(':').pop().split('/').pop()
+      return cluster.split(':').pop()?.split('/').pop()
     }
     return cluster
   }
-  if (cluster.GetAtt && cluster.GetAtt[1] === 'Arn') {
+  if (cluster.GetAtt != null && cluster.GetAtt[1] === 'Arn') {
     // AWS::ECS::Cluster returns the cluster name for 'Ref'
     return { Ref: cluster.GetAtt[0] }
   }
   return cluster // Fallback to name
 }
 
+const executionMetrics: EcsMEtrics[] = ['MemoryUtilization', 'CPUUtilization']
+
 /**
  * ecsAlarmsConfig The fully resolved alarm configuration
  */
-export default function ecsAlarms (ecsAlarmsConfig: EcsAlarmsConfig, context: Context) {
-  return {
-    createECSAlarms
-  }
-
+export default function createECSAlarms (ecsAlarmsConfig: EcsAlarmsConfig, context: Context, compiledTemplate: Template, additionalResources: ResourceType = {}) {
   /**
    * Add all required ECS alarms to the provided CloudFormation template
    * based on the ECS Service resources
    *
    * A CloudFormation template object
    */
-  function createECSAlarms (cfTemplate: CloudFormationTemplate) {
-    const serviceResources = cfTemplate.getResourcesByType(
-      'AWS::ECS::Service'
-    )
-    for (const [serviceResourceName, serviceResource] of Object.entries(
-      serviceResources
-    )) {
-      const cluster = serviceResource.Properties.Cluster
+  const serviceResources = getResourcesByType('AWS::ECS::Service', compiledTemplate, additionalResources)
+
+  for (const [serviceResourceName, serviceResource] of Object.entries(serviceResources)) {
+    for (const metric of executionMetrics) {
+      const cluster = serviceResource.Properties?.Cluster
       const clusterName = resolveEcsClusterNameAsCfn(cluster)
-      if (ecsAlarmsConfig.MemoryUtilization.ActionsEnabled) {
-        const memoryUtilizationAlarm = createMemoryUtilizationAlarm(
-          serviceResourceName,
-          serviceResource,
-          clusterName,
-          ecsAlarmsConfig.MemoryUtilization
-        )
-        cfTemplate.addResource(memoryUtilizationAlarm.resourceName, memoryUtilizationAlarm.resource)
+      const config = ecsAlarmsConfig[metric]
+      if (config.enabled !== false) {
+        const threshold = config.Threshold
+        const ecsAlarmProperties: EcsAlarm = {
+          AlarmName: `ECS_${metric.replaceAll('Utilization', 'Alarm')}_\${${serviceResourceName}.Name}`,
+          AlarmDescription: `ECS ${metric} for ${serviceResourceName}.Name breaches ${threshold}`,
+          ServiceName: `\${${serviceResourceName}.Name}`,
+          ClusterName: clusterName,
+          MetricName: metric,
+          Namespace: 'AWS/ECS',
+          Dimensions: [
+            { Name: 'ServiceName', Value: `\${${serviceResourceName}.Name}` },
+            { Name: 'ClusterName', Value: clusterName }
+          ],
+          ...config
+        }
+        const resourceName = `slicWatch${metric}Alarm${serviceResourceName}`
+        const resource = createAlarm(ecsAlarmProperties, context)
+        addResource(resourceName, resource, compiledTemplate)
       }
-      if (ecsAlarmsConfig.CPUUtilization.ActionsEnabled) {
-        const cpuUtilizationAlarm = createCPUUtilizationAlarm(
-          serviceResourceName,
-          serviceResource,
-          clusterName,
-          ecsAlarmsConfig.CPUUtilization
-        )
-        cfTemplate.addResource(cpuUtilizationAlarm.resourceName, cpuUtilizationAlarm.resource)
-      }
-    }
-  }
-
-  function createMemoryUtilizationAlarm (logicalId: string, serviceResource: Resource, clusterName: string, config: AlarmProperties) {
-    const threshold = config.Threshold
-    const ecsAlarmProperties: EcsAlarm = {
-      AlarmName: `ECS_MemoryAlarm_\${${logicalId}.Name}`,
-      AlarmDescription: `ECS memory utilization for ${logicalId}.Name breaches ${threshold}`,
-      ServiceName: `\${${logicalId}.Name}`,
-      ClusterName: clusterName,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: 'MemoryUtilization',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/ECS',
-      Dimensions: [
-        { Name: 'ServiceName', Value: `\${${logicalId}.Name}` },
-        { Name: 'ClusterName', Value: clusterName }
-      ]
-    }
-    return {
-      resourceName: `slicWatchECSMemoryAlarm${logicalId}`,
-      resource: createAlarm(ecsAlarmProperties, context)
-    }
-  }
-
-  function createCPUUtilizationAlarm (logicalId: string, serviceResource: Resource, clusterName: string, config: AlarmProperties) {
-    const threshold = config.Threshold
-    const ecsAlarmProperties: EcsAlarm = {
-      AlarmName: `ECS_CPUAlarm_\${${logicalId}.Name}`,
-      AlarmDescription: `ECS CPU utilization for ${logicalId}.Name breaches ${threshold}`,
-      ServiceName: `\${${logicalId}.Name}`,
-      ClusterName: clusterName,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: 'CPUUtilization',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/ECS',
-      Dimensions: [
-        { Name: 'ServiceName', Value: `\${${logicalId}.Name}` },
-        { Name: 'ClusterName', Value: clusterName }
-      ]
-    }
-    return {
-      resourceName: `slicWatchECSCPUAlarm${logicalId}`,
-      resource: createAlarm(ecsAlarmProperties, context)
     }
   }
 }
