@@ -1,102 +1,55 @@
+
 'use strict'
 
-import { CloudFormationTemplate } from '../cf-template'
-import Resource from 'cloudform-types/types/resource'
-import { Context, createAlarm } from './default-config-alarms'
-import { AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import { getResourcesByType, addResource } from '../cf-template'
+import { type Context, createAlarm, type DefaultAlarmsProperties } from './default-config-alarms'
+import { type AlarmProperties } from 'cloudform-types/types/cloudWatch/alarm'
+import type Template from 'cloudform-types/types/template'
+import { type ResourceType } from './../cf-template'
 
-export type EventsAlarmsConfig = AlarmProperties &{
-  FailedInvocations: AlarmProperties,
-  ThrottledRules: AlarmProperties
+export interface EventsAlarmsConfig {
+  enabled?: boolean
+  FailedInvocations: DefaultAlarmsProperties
+  ThrottledRules: DefaultAlarmsProperties
 }
 
 export type EventbridgeAlarm = AlarmProperties & {
   RuleName: string
 }
 
+type EventMetrics = 'FailedInvocations' | 'ThrottledRules'
+
+const executionMetrics: EventMetrics[] = ['FailedInvocations', 'ThrottledRules']
+
 /**
  * The fully resolved alarm configuration
  */
-export default function eventsAlarms (eventsAlarmsConfig: EventsAlarmsConfig, context: Context) {
-  return {
-    createRuleAlarms
-  }
-
+export default function createRuleAlarms (eventsAlarmsConfig: EventsAlarmsConfig, context: Context, compiledTemplate: Template, additionalResources: ResourceType = {}) {
   /**
    * Add all required Events alarms to the provided CloudFormation template
    * based on the EventBridge Rule found within
    *
-   * cfTemplate A CloudFormation template object
    */
-  function createRuleAlarms (cfTemplate:CloudFormationTemplate) {
-    const ruleResources = cfTemplate.getResourcesByType(
-      'AWS::Events::Rule'
-    )
+  const ruleResources = getResourcesByType('AWS::Events::Rule', compiledTemplate, additionalResources)
 
-    for (const [ruleResourceName, ruleResource] of Object.entries(ruleResources)) {
-      if (eventsAlarmsConfig.FailedInvocations.ActionsEnabled) {
-        const failedInvocations = createFailedInvocationsAlarm(
-          ruleResourceName,
-          ruleResource,
-          eventsAlarmsConfig.FailedInvocations
-        )
-        cfTemplate.addResource(failedInvocations.resourceName, failedInvocations.resource)
+  for (const [ruleResourceName] of Object.entries(ruleResources)) {
+    for (const metric of executionMetrics) {
+      const config: DefaultAlarmsProperties = eventsAlarmsConfig[metric]
+      if (eventsAlarmsConfig[metric].enabled !== false) {
+        const threshold = config.Threshold
+        const eventbridgeAlarmProperties: EventbridgeAlarm = {
+          AlarmName: `Events_${metric}Alarm_${ruleResourceName}`,
+          AlarmDescription: `EventBridge ${metric} for \${${ruleResourceName}} breaches ${threshold}`,
+          RuleName: `${ruleResourceName}`,
+          MetricName: metric,
+          Namespace: 'AWS/Events',
+          Dimensions: [{ Name: 'RuleName', Value: { Ref: ruleResourceName } as any }],
+          ...config
+        }
+        const resourceName = `slicWatchEvents${metric}Alarm${ruleResourceName}`
+        const resource = createAlarm(eventbridgeAlarmProperties, context)
+        addResource(resourceName, resource, compiledTemplate)
       }
-
-      if (eventsAlarmsConfig.ThrottledRules.ActionsEnabled) {
-        const throttledRules = createThrottledRulesAlarm(
-          ruleResourceName,
-          ruleResource,
-          eventsAlarmsConfig.ThrottledRules
-        )
-        cfTemplate.addResource(throttledRules.resourceName, throttledRules.resource)
-      }
-    }
-  }
-
-  function createFailedInvocationsAlarm (logicalId: string, ruleResource: Resource, config: AlarmProperties) {
-    const threshold = config.Threshold
-    const eventbridgeAlarmProperties:EventbridgeAlarm = {
-      AlarmName: `Events_FailedInvocationsAlarm_${logicalId}`,
-      AlarmDescription: `EventBridge Failed Invocations for \${${logicalId}} breaches ${threshold}`,
-      RuleName: `${logicalId}`,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: 'FailedInvocations',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/Events',
-      Dimensions: [{ Name: 'RuleName', Value: { Ref: logicalId } as any }]
-    }
-    return {
-      resourceName: `slicWatchEventsFailedInvocationsAlarm${logicalId}`,
-      resource: createAlarm(eventbridgeAlarmProperties, context)
-    }
-  }
-
-  function createThrottledRulesAlarm (logicalId: string, ruleResource: Resource, config: AlarmProperties) {
-    const threshold = config.Threshold
-    const eventbridgeAlarmProperties:EventbridgeAlarm = {
-      AlarmName: `Events_ThrottledRulesAlarm_${logicalId}`,
-      AlarmDescription: `EventBridge Throttled Rules for ${logicalId} breaches ${threshold}`,
-      RuleName: `${logicalId}`,
-      ComparisonOperator: config.ComparisonOperator,
-      Threshold: config.Threshold,
-      MetricName: 'ThrottledRules',
-      Statistic: config.Statistic,
-      Period: config.Period,
-      ExtendedStatistic: config.ExtendedStatistic,
-      EvaluationPeriods: config.EvaluationPeriods,
-      TreatMissingData: config.TreatMissingData,
-      Namespace: 'AWS/Events',
-      Dimensions: [{ Name: 'RuleName', Value: { Ref: logicalId } as any }]
-    }
-    return {
-      resourceName: `slicWatchEventsThrottledRulesAlarm${logicalId}`,
-      resource: createAlarm(eventbridgeAlarmProperties, context)
     }
   }
 }
