@@ -1,10 +1,12 @@
 'use strict'
 
+import { type AllAlarms } from './default-config-alarms'
+
 /*
  * Determine the presentation name for an alarm statistic
  *
  * @param {*} AlarmProperties Alarm configuration
  */
-export function getStatisticName (AlarmProperties) {
-  return AlarmProperties.Statistic || AlarmProperties.ExtendedStatistic
+export function getStatisticName (AlarmProperties: AllAlarms) {
+  return AlarmProperties.Statistic ?? AlarmProperties.ExtendedStatistic
 }
