@@ -2,23 +2,23 @@
 
 import { test } from 'tap'
 
-import alarms from '../alarms'
+import addAlarms from '../alarms'
 import defaultConfig from '../../inputs/default-config'
 import { createTestCloudFormationTemplate, albCfTemplate, createTestConfig, testContext } from '../../tests/testing-utils'
+import { getResourcesByType } from '../../cf-template'
 
 test('Alarms create all service alarms', (t) => {
-  const cfTemplate = createTestCloudFormationTemplate()
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
   const funcAlarmPropertiess = {}
-  for (const funcLogicalId of Object.keys(cfTemplate.getResourcesByType('AWS::Lambda::Function'))) {
+  for (const funcLogicalId of Object.keys(getResourcesByType('AWS::Lambda::Function', compiledTemplate))) {
     funcAlarmPropertiess[funcLogicalId] = {}
   }
-  const { addAlarms } = alarms(defaultConfig.alarms, funcAlarmPropertiess, testContext)
-  addAlarms(cfTemplate)
+  addAlarms(defaultConfig.alarms, funcAlarmPropertiess, testContext, compiledTemplate, additionalResources)
   const namespaces = new Set()
   for (const resource of Object.values(
-    cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+    getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
   )) {
-    if (resource.Properties.Namespace) {
+    if (resource.Properties?.Namespace != null) {
       namespaces.add(resource.Properties.Namespace)
     }
   }
@@ -27,18 +27,17 @@ test('Alarms create all service alarms', (t) => {
 })
 
 test('Alarms create all ALB service alarms', (t) => {
-  const cfTemplate = createTestCloudFormationTemplate(albCfTemplate)
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate(albCfTemplate)
   const funcAlarmPropertiess = {}
-  for (const funcLogicalId of Object.keys(cfTemplate.getResourcesByType('AWS::Lambda::Function'))) {
+  for (const funcLogicalId of Object.keys(getResourcesByType('AWS::Lambda::Function', compiledTemplate))) {
     funcAlarmPropertiess[funcLogicalId] = {}
   }
-  const { addAlarms } = alarms(defaultConfig.alarms, funcAlarmPropertiess, testContext)
-  addAlarms(cfTemplate)
+  addAlarms(defaultConfig.alarms, funcAlarmPropertiess, testContext, compiledTemplate, additionalResources)
   const namespaces = new Set()
   for (const resource of Object.values(
-    cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+    getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
   )) {
-    if (resource.Properties.Namespace) {
+    if (resource.Properties?.Namespace != null) {
       namespaces.add(resource.Properties.Namespace)
     }
   }
@@ -50,18 +49,17 @@ test('Alarms are not created when disabled globally', (t) => {
   const config = createTestConfig(
     defaultConfig.alarms,
     {
-      ActionsEnabled: false
+      enabled: false
     }
   )
-  const cfTemplate = createTestCloudFormationTemplate()
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
   const funcAlarmPropertiess = {}
-  for (const funcLogicalId of Object.keys(cfTemplate.getResourcesByType('AWS::Lambda::Function'))) {
+  for (const funcLogicalId of Object.keys(getResourcesByType('AWS::Lambda::Function', compiledTemplate))) {
     funcAlarmPropertiess[funcLogicalId] = {}
   }
-  const { addAlarms } = alarms(config, funcAlarmPropertiess, testContext)
-  addAlarms(cfTemplate)
+  addAlarms(config, funcAlarmPropertiess, testContext, compiledTemplate, additionalResources)
 
-  const alarmsCreated = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmsCreated = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmsCreated)
   t.end()
