@@ -1,7 +1,6 @@
-/* eslint-disable no-template-curly-in-string */
 'use strict'
 
-import albTargetAlarms, { AlbTargetAlarmProperties, findLoadBalancersForTargetGroup } from '../alb-target-group'
+import createALBTargetAlarms, { findLoadBalancersForTargetGroup } from '../alb-target-group'
 import { test } from 'tap'
 import { defaultConfig } from '../../inputs/default-config'
 import {
@@ -12,26 +11,24 @@ import {
   albCfTemplate,
   testContext
 } from '../../tests/testing-utils'
-
-import CloudFormationTemplate from '../../cf-template'
-const albCfTemp = CloudFormationTemplate(albCfTemplate)
+import { getResourcesByType } from '../../cf-template'
 
 test('findLoadBalancersForTargetGroup', (t) => {
   test('finds the associated Load Balancer if it exists in the CloudFormation template for the Target Group', (t) => {
     const targetGroupLogicalId = 'AlbEventAlbTargetGrouphttpListener'
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup(targetGroupLogicalId, albCfTemp)
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup(targetGroupLogicalId, albCfTemplate, {})
     t.same(loadBalancerLogicalIds, ['alb'])
     t.end()
   })
 
   test('returns empty list for non-existent listener', (t) => {
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('fakeListener', CloudFormationTemplate({}))
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('fakeListener', {}, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
 
   test('includes an ALB from the DefaultActions', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listener: {
           Type: 'AWS::ElasticLoadBalancingV2::Listener',
@@ -49,15 +46,15 @@ test('findLoadBalancersForTargetGroup', (t) => {
         },
         alb: { Type: '' }
       }
-    })
+    }
 
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', template)
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', compiledTemplate, {})
     t.same(loadBalancerLogicalIds, ['alb'])
     t.end()
   })
 
   test('excludes DefaultActions with a literal load balancer ARN', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listener: {
           Type: 'AWS::ElasticLoadBalancingV2::Listener',
@@ -75,14 +72,14 @@ test('findLoadBalancersForTargetGroup', (t) => {
         },
         alb: { Type: '' }
       }
-    })
+    }
 
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', template)
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', compiledTemplate, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
   test('finds load balancers through listener rule target groups', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listenerRuleA: {
           Type: 'AWS::ElasticLoadBalancingV2::ListenerRule',
@@ -105,14 +102,14 @@ test('findLoadBalancersForTargetGroup', (t) => {
           }
         }
       }
-    })
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', template)
+    }
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', compiledTemplate, {})
     t.same(loadBalancerLogicalIds, ['alb'])
     t.end()
   })
 
   test('omits listener rules with no load balancer', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listenerRule: {
           Type: 'AWS::ElasticLoadBalancingV2::ListenerRule',
@@ -126,14 +123,14 @@ test('findLoadBalancersForTargetGroup', (t) => {
           Properties: {}
         }
       }
-    })
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', template)
+    }
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', compiledTemplate, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
 
   test('omits listener rules with no actions', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listenerRule: {
           Type: 'AWS::ElasticLoadBalancingV2::ListenerRule',
@@ -142,14 +139,14 @@ test('findLoadBalancersForTargetGroup', (t) => {
           }
         }
       }
-    })
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', template)
+    }
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tg', compiledTemplate, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
 
   test('omits listener rules a literal listener ARN', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listenerRule: {
           Type: 'AWS::ElasticLoadBalancingV2::ListenerRule',
@@ -159,14 +156,14 @@ test('findLoadBalancersForTargetGroup', (t) => {
           }
         }
       }
-    })
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', template)
+    }
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', compiledTemplate, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
 
   test('omits listeners with a literal load balancer ARN', (t) => {
-    const template = CloudFormationTemplate({
+    const compiledTemplate = {
       Resources: {
         listenerRuleA: {
           Type: 'AWS::ElasticLoadBalancingV2::ListenerRule',
@@ -182,8 +179,8 @@ test('findLoadBalancersForTargetGroup', (t) => {
           }
         }
       }
-    })
-    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', template)
+    }
+    const loadBalancerLogicalIds = findLoadBalancersForTargetGroup('tgA', compiledTemplate, {})
     t.equal(loadBalancerLogicalIds.length, 0)
     t.end()
   })
@@ -216,13 +213,12 @@ test('ALB Target Group alarms are created', (t) => {
     }
 
   )
-  function createAlarmResources (elbAlarmProperties:AlbTargetAlarmProperties) {
-    const { createALBTargetAlarms } = albTargetAlarms(elbAlarmProperties, testContext)
-    const cfTemplate = createTestCloudFormationTemplate(albCfTemplate)
-    createALBTargetAlarms(cfTemplate)
-    return cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
-  }
-  const targeGroupAlarmResources = createAlarmResources(AlarmPropertiesTargetGroup.ApplicationELBTarget)
+
+  const albAlarmProperties = AlarmPropertiesTargetGroup.ApplicationELBTarget
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate(albCfTemplate)
+  createALBTargetAlarms(albAlarmProperties, testContext, compiledTemplate, additionalResources)
+
+  const targetGroupAlarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   const expectedTypesTargetGroup = {
     LoadBalancerHTTPCodeTarget5XXCountAlarm: 'HTTPCode_Target_5XX_Count',
@@ -231,21 +227,21 @@ test('ALB Target Group alarms are created', (t) => {
     LoadBalancerLambdaUserErrorAlarm: 'LambdaUserError'
   }
 
-  t.equal(Object.keys(targeGroupAlarmResources).length, Object.keys(expectedTypesTargetGroup).length)
-  for (const alarmResource of Object.values(targeGroupAlarmResources)) {
+  t.equal(Object.keys(targetGroupAlarmResources).length, Object.keys(expectedTypesTargetGroup).length)
+  for (const alarmResource of Object.values(targetGroupAlarmResources)) {
     const al = alarmResource.Properties
     assertCommonAlarmProperties(t, al)
-    const alarmType = alarmNameToType(al.AlarmName)
+    const alarmType = alarmNameToType(al?.AlarmName)
     const expectedMetric = expectedTypesTargetGroup[alarmType]
-    t.equal(al.MetricName, expectedMetric)
-    t.ok(al.Statistic)
-    t.equal(al.Threshold, AlarmPropertiesTargetGroup.ApplicationELBTarget[expectedMetric].Threshold)
-    t.equal(al.EvaluationPeriods, 2)
-    t.equal(al.TreatMissingData, 'breaching')
-    t.equal(al.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
-    t.equal(al.Namespace, 'AWS/ApplicationELB')
-    t.equal(al.Period, 120)
-    t.same(al.Dimensions, [
+    t.equal(al?.MetricName, expectedMetric)
+    t.ok(al?.Statistic)
+    t.equal(al?.Threshold, albAlarmProperties[expectedMetric].Threshold)
+    t.equal(al?.EvaluationPeriods, 2)
+    t.equal(al?.TreatMissingData, 'breaching')
+    t.equal(al?.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
+    t.equal(al?.Namespace, 'AWS/ApplicationELB')
+    t.equal(al?.Period, 120)
+    t.same(al?.Dimensions, [
       {
         Name: 'TargetGroup',
         Value: '${AlbEventAlbTargetGrouphttpListener.TargetGroupFullName}'
@@ -265,7 +261,7 @@ test('ALB alarms are not created when disabled globally', (t) => {
     defaultConfig.alarms,
     {
       ApplicationELBTarget: {
-        ActionsEnabled: false, // disabled globally
+        enabled: false, // disabled globally
         Period: 60,
         HTTPCode_Target_5XX_Count: {
           Threshold: 50
@@ -283,14 +279,11 @@ test('ALB alarms are not created when disabled globally', (t) => {
     }
   )
 
-  function createAlarmResources (elbAlarmProperties:AlbTargetAlarmProperties) {
-    const { createALBTargetAlarms } = albTargetAlarms(elbAlarmProperties, testContext)
-    const cfTemplate = createTestCloudFormationTemplate(albCfTemplate)
-    createALBTargetAlarms(cfTemplate)
-    return cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
-  }
-  const targeGroupAlarmResources = createAlarmResources(AlarmPropertiesTargetGroup.ApplicationELBTarget)
+  const albAlarmProperties = AlarmPropertiesTargetGroup.ApplicationELBTarget
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate(albCfTemplate)
+  createALBTargetAlarms(albAlarmProperties, testContext, compiledTemplate, additionalResources)
+  const targetGroupAlarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
-  t.same({}, targeGroupAlarmResources)
+  t.same({}, targetGroupAlarmResources)
   t.end()
 })
