@@ -1,6 +1,7 @@
 'use strict'
 
-import ruleAlarms from '../eventbridge'
+import createRuleAlarms from '../eventbridge'
+import { getResourcesByType } from '../../cf-template'
 import { test } from 'tap'
 import defaultConfig from '../../inputs/default-config'
 import {
@@ -30,12 +31,10 @@ test('Events alarms are created', (t) => {
     }
   )
   const ruleAlarmProperties = AlarmProperties.Events
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createRuleAlarms(ruleAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createRuleAlarms } = ruleAlarms(ruleAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createRuleAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   const expectedTypes = {
     Events_FailedInvocationsAlarm: 'FailedInvocations',
@@ -46,22 +45,19 @@ test('Events alarms are created', (t) => {
   for (const alarmResource of Object.values(alarmResources)) {
     const al = alarmResource.Properties
     assertCommonAlarmProperties(t, al)
-    const alarmType = alarmNameToType(al.AlarmName)
+    const alarmType = alarmNameToType(al?.AlarmName)
     const expectedMetric = expectedTypes[alarmType]
-    t.equal(al.MetricName, expectedMetric)
-    t.ok(al.Statistic)
-    t.equal(al.Threshold, ruleAlarmProperties[expectedMetric].Threshold)
-    t.equal(al.EvaluationPeriods, 2)
-    t.equal(al.TreatMissingData, 'breaching')
-    t.equal(al.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
-    t.equal(al.Namespace, 'AWS/Events')
-    t.equal(al.Period, 120)
-    // @ts-ignore
-    t.equal(al.Dimensions.length, 1)
-    // @ts-ignore
-    t.equal(al.Dimensions[0].Name, 'RuleName')
-    // @ts-ignore
-    t.ok(al.Dimensions[0].Value)
+    t.equal(al?.MetricName, expectedMetric)
+    t.ok(al?.Statistic)
+    t.equal(al?.Threshold, ruleAlarmProperties[expectedMetric].Threshold)
+    t.equal(al?.EvaluationPeriods, 2)
+    t.equal(al?.TreatMissingData, 'breaching')
+    t.equal(al?.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
+    t.equal(al?.Namespace, 'AWS/Events')
+    t.equal(al?.Period, 120)
+    t.equal(al?.Dimensions.length, 1)
+    t.equal(al?.Dimensions[0].Name, 'RuleName')
+    t.ok(al?.Dimensions[0].Value)
   }
 
   t.end()
@@ -72,7 +68,7 @@ test('Events alarms are not created when disabled globally', (t) => {
     defaultConfig.alarms,
     {
       Events: {
-        ActionsEnabled: false, // disabled globally
+        enabled: false, // disabled globally
         Period: 60,
         FailedInvocations: {
           Threshold: 50
@@ -84,13 +80,10 @@ test('Events alarms are not created when disabled globally', (t) => {
     }
   )
   const ruleAlarmProperties = AlarmProperties.Events
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createRuleAlarms(ruleAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createRuleAlarms } = ruleAlarms(ruleAlarmProperties, testContext)
-
-  const cfTemplate = createTestCloudFormationTemplate()
-  createRuleAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmResources)
   t.end()
