@@ -1,6 +1,7 @@
 'use strict'
 
-import kinesisAlarms from '../kinesis'
+import createKinesisAlarms from '../kinesis'
+import { getResourcesByType } from '../../cf-template'
 import { test } from 'tap'
 import defaultConfig from '../../inputs/default-config'
 import {
@@ -27,12 +28,10 @@ test('Kinesis data stream alarms are created', (t) => {
     }
   )
   const kinesisAlarmProperties = AlarmProperties.Kinesis
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createKinesisAlarms(kinesisAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createKinesisAlarms } = kinesisAlarms(kinesisAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createKinesisAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   const expectedTypes = {
     Kinesis_StreamIteratorAge: 'GetRecords.IteratorAgeMilliseconds',
@@ -47,17 +46,17 @@ test('Kinesis data stream alarms are created', (t) => {
   for (const alarmResource of Object.values(alarmResources)) {
     const al = alarmResource.Properties
     assertCommonAlarmProperties(t, al)
-    const alarmType = alarmNameToType(al.AlarmName)
+    const alarmType = alarmNameToType(al?.AlarmName)
     const expectedMetric = expectedTypes[alarmType]
-    t.equal(al.MetricName, expectedMetric)
-    t.ok(al.Statistic)
-    t.equal(al.Threshold, kinesisAlarmProperties[expectedMetric].Threshold)
-    t.equal(al.EvaluationPeriods, 2)
-    t.equal(al.TreatMissingData, 'breaching')
-    t.equal(al.ComparisonOperator, 'LessThanThreshold')
-    t.equal(al.Namespace, 'AWS/Kinesis')
-    t.equal(al.Period, 120)
-    t.same(al.Dimensions, [
+    t.equal(al?.MetricName, expectedMetric)
+    t.ok(al?.Statistic)
+    t.equal(al?.Threshold, kinesisAlarmProperties[expectedMetric].Threshold)
+    t.equal(al?.EvaluationPeriods, 2)
+    t.equal(al?.TreatMissingData, 'breaching')
+    t.equal(al?.ComparisonOperator, 'LessThanThreshold')
+    t.equal(al?.Namespace, 'AWS/Kinesis')
+    t.equal(al?.Period, 120)
+    t.same(al?.Dimensions, [
       {
         Name: 'StreamName',
         Value: {
@@ -75,7 +74,7 @@ test('Kinesis data stream alarms are not created when disabled globally', (t) =>
     defaultConfig.alarms,
     {
       Kinesis: {
-        ActionsEnabled: false, // disabled globally
+        enabled: false, // disabled globally
         Period: 60,
         IteratorAgeMilliseconds: {
           Threshold: 5000
@@ -84,13 +83,10 @@ test('Kinesis data stream alarms are not created when disabled globally', (t) =>
     }
   )
   const kinesisAlarmProperties = AlarmProperties.Kinesis
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createKinesisAlarms(kinesisAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createKinesisAlarms } = kinesisAlarms(kinesisAlarmProperties, testContext)
-
-  const cfTemplate = createTestCloudFormationTemplate()
-  createKinesisAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmResources)
   t.end()
