@@ -1,7 +1,7 @@
-/* eslint-disable no-template-curly-in-string */
 'use strict'
 
-import snsAlarms from '../sns'
+import createSNSAlarms from '../sns'
+import { getResourcesByType } from '../../cf-template'
 import { test } from 'tap'
 import defaultConfig from '../../inputs/default-config'
 import {
@@ -31,12 +31,10 @@ test('SNS alarms are created', (t) => {
     }
   )
   const snsAlarmProperties = AlarmProperties.SNS
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createSNSAlarms(snsAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createSNSAlarms } = snsAlarms(snsAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createSNSAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   const expectedTypes = {
     SNS_NumberOfNotificationsFilteredOutInvalidAttributesAlarm: 'NumberOfNotificationsFilteredOut-InvalidAttributes',
@@ -47,17 +45,17 @@ test('SNS alarms are created', (t) => {
   for (const alarmResource of Object.values(alarmResources)) {
     const al = alarmResource.Properties
     assertCommonAlarmProperties(t, al)
-    const alarmType = alarmNameToType(al.AlarmName)
+    const alarmType = alarmNameToType(al?.AlarmName)
     const expectedMetric = expectedTypes[alarmType]
-    t.equal(al.MetricName, expectedMetric)
-    t.ok(al.Statistic)
-    t.equal(al.Threshold, snsAlarmProperties[expectedMetric].Threshold)
-    t.equal(al.EvaluationPeriods, 2)
-    t.equal(al.TreatMissingData, 'breaching')
-    t.equal(al.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
-    t.equal(al.Namespace, 'AWS/SNS')
-    t.equal(al.Period, 120)
-    t.same(al.Dimensions, [
+    t.equal(al?.MetricName, expectedMetric)
+    t.ok(al?.Statistic)
+    t.equal(al?.Threshold, snsAlarmProperties[expectedMetric].Threshold)
+    t.equal(al?.EvaluationPeriods, 2)
+    t.equal(al?.TreatMissingData, 'breaching')
+    t.equal(al?.ComparisonOperator, 'GreaterThanOrEqualToThreshold')
+    t.equal(al?.Namespace, 'AWS/SNS')
+    t.equal(al?.Period, 120)
+    t.same(al?.Dimensions, [
       {
         Name: 'TopicName',
         Value: '${topic.TopicName}'
@@ -73,7 +71,7 @@ test('SNS alarms are not created when disabled globally', (t) => {
     defaultConfig.alarms,
     {
       SNS: {
-        ActionsEnabled: false, // disabled globally
+        enabled: false, // disabled globally
         Period: 60,
         'NumberOfNotificationsFilteredOut-InvalidAttributes': {
           Threshold: 50
@@ -85,13 +83,10 @@ test('SNS alarms are not created when disabled globally', (t) => {
     }
   )
   const snsAlarmProperties = AlarmProperties.SNS
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createSNSAlarms(snsAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createSNSAlarms } = snsAlarms(snsAlarmProperties, testContext)
-
-  const cfTemplate = createTestCloudFormationTemplate()
-  createSNSAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmResources)
   t.end()
