@@ -1,6 +1,7 @@
 'use strict'
 
-import stepFunctionsAlarms from '../step-functions'
+import createStatesAlarms from '../step-functions'
+import { getResourcesByType } from '../../cf-template'
 import { test } from 'tap'
 import defaultConfig from '../../inputs/default-config'
 import {
@@ -34,12 +35,9 @@ test('Step Function alarms are created', (t) => {
     }
   )
   const sfAlarmProperties = AlarmProperties.States
-
-  const { createStatesAlarms } = stepFunctionsAlarms(sfAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createStatesAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createStatesAlarms(sfAlarmProperties, testContext, compiledTemplate, additionalResources)
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   const alarmsByType = {}
   t.equal(Object.keys(alarmResources).length, 6)
@@ -48,8 +46,8 @@ test('Step Function alarms are created', (t) => {
     if (!resourceName.endsWith('ExpressWorkflow')) {
       const al = alarmResource.Properties
       assertCommonAlarmProperties(t, al)
-      const alarmType = alarmNameToType(al.AlarmName)
-      alarmsByType[alarmType] = alarmsByType[alarmType] || new Set()
+      const alarmType = alarmNameToType(al?.AlarmName)
+      alarmsByType[alarmType] = alarmsByType[alarmType] ?? new Set()
       alarmsByType[alarmType].add(al)
     }
   }
@@ -92,7 +90,7 @@ test('Step function alarms are not created when disabled globally', (t) => {
     defaultConfig.alarms,
     {
       States: {
-        ActionsEnabled: false, // disabled globally
+        enabled: false, // disabled globally
         Period: 900,
         ExecutionThrottled: {
           Threshold: 0
@@ -107,12 +105,10 @@ test('Step function alarms are not created when disabled globally', (t) => {
     }
   )
   const sfAlarmProperties = AlarmProperties.States
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createStatesAlarms(sfAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createStatesAlarms } = stepFunctionsAlarms(sfAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createStatesAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmResources)
   t.end()
@@ -123,30 +119,28 @@ test('Step function alarms are not created when disabled individually', (t) => {
     defaultConfig.alarms,
     {
       States: {
-        ActionsEnabled: true, // enabdled globally
+        enabled: true, // enabdled globally
         Period: 900,
         ExecutionThrottled: {
-          ActionsEnabled: false, // disabled locally
+          enabled: false, // disabled locally
           Threshold: 0
         },
         ExecutionsFailed: {
-          ActionsEnabled: false, // disabled locally
+          enabled: false, // disabled locally
           Threshold: 0
         },
         ExecutionsTimedOut: {
-          ActionsEnabled: false, // disabled locally
+          enabled: false, // disabled locally
           Threshold: 0
         }
       }
     }
   )
   const sfAlarmProperties = AlarmProperties.States
+  const { compiledTemplate, additionalResources } = createTestCloudFormationTemplate()
+  createStatesAlarms(sfAlarmProperties, testContext, compiledTemplate, additionalResources)
 
-  const { createStatesAlarms } = stepFunctionsAlarms(sfAlarmProperties, testContext)
-  const cfTemplate = createTestCloudFormationTemplate()
-  createStatesAlarms(cfTemplate)
-
-  const alarmResources = cfTemplate.getResourcesByType('AWS::CloudWatch::Alarm')
+  const alarmResources = getResourcesByType('AWS::CloudWatch::Alarm', compiledTemplate)
 
   t.same({}, alarmResources)
   t.end()
