@@ -1,4 +1,5 @@
 'use strict'
+
 import Ajv from 'ajv'
 import { test } from 'tap'
 import defaultConfig from '../default-config'
@@ -15,12 +16,10 @@ try {
     })
     const slicWatchValidate = ajv.compile(slicWatchSchema)
     const slicWatchValid = slicWatchValidate(slicWatchConfig)
-    // @ts-ignore
     t.ok(slicWatchValid, slicWatchValidate.errors)
     const pluginValidate = ajv.compile(pluginConfigSchema)
     const testConfig = { slicWatch: slicWatchConfig }
     const pluginValid = pluginValidate(testConfig)
-    // @ts-ignore
     t.ok(pluginValid, pluginValidate.errors)
     t.end()
   })
@@ -38,13 +37,11 @@ test('Default config conforms to the config schema without topicArn', (t) => {
   })
   const slicWatchValidate = ajv.compile(slicWatchSchema)
   const slicWatchValid = slicWatchValidate(slicWatchConfig)
-  // @ts-ignore
   t.ok(slicWatchValid, slicWatchValidate.errors)
 
   const pluginValidate = ajv.compile(pluginConfigSchema)
   const testConfig = { slicWatch: slicWatchConfig }
   const pluginValid = pluginValidate(testConfig)
-  // @ts-ignore
   t.ok(pluginValid, pluginValidate.errors)
 
   t.end()
