@@ -1,22 +1,25 @@
+
 'use strict'
 
 import _ from 'lodash'
 import Ajv from 'ajv'
-
-import alarms from '../core/alarms/alarms'
-import dashboard from '../core/dashboards/dashboard'
+import addAlarms from '../core/alarms/alarms'
+import addDashboard from '../core/dashboards/dashboard'
 import { pluginConfigSchema, functionConfigSchema, slicWatchSchema } from '../core/inputs/config-schema'
 import defaultConfig from '../core/inputs/default-config'
-import CloudFormationTemplate from '../core/cf-template'
 import Serverless from 'serverless'
-import Hooks from 'serverless-hooks-plugin'
-import Aws from 'serverless/plugins/aws/provider/awsProvider'
+import type Hooks from 'serverless-hooks-plugin'
+import type Aws from 'serverless/plugins/aws/provider/awsProvider'
+import { type ResourceType } from './../core/cf-template'
+
+interface SlicWatchConfig {
+  topicArn?: string
+  enabled?: boolean
+}
 class ServerlessPlugin {
   serverless: Serverless
   hooks: Hooks
   providerNaming: Aws
-  dashboard: { addDashboard: (cfTemplate: import('slic-watch-core/cf-template').CloudFormationTemplate) => void; }
-  alarms: { addAlarms: (cfTemplate: import('slic-watch-core/cf-template').CloudFormationTemplate) => void; }
 
   /**
    * Plugin constructor according to the Serverless Framework v2/v3 plugin signature
@@ -28,10 +31,11 @@ class ServerlessPlugin {
     this.providerNaming = serverless.providers.aws.naming
 
     if (serverless.service.provider.name !== 'aws') {
+      // eslint-disable-next-line @typescript-eslint/no-throw-literal
       throw new Serverless('SLIC Watch only supports AWS')
     }
 
-    if (serverless.configSchemaHandler) {
+    if (serverless.configSchemaHandler != null) {
       serverless.configSchemaHandler.defineCustomProperties(pluginConfigSchema)
       serverless.configSchemaHandler.defineFunctionProperties('aws', functionConfigSchema)
     }
@@ -43,13 +47,8 @@ class ServerlessPlugin {
   /**
    * Modify the CloudFormation template before the package is finalized
    */
-  createSlicWatchResources () {
-    type SlicWatchConfig = {
-      topicArn: string
-      ActionsEnabled: boolean
-    }
-
-    const slicWatchConfig: SlicWatchConfig = (this.serverless.service.custom || {}).slicWatch || {}
+  createSlicWatchResources (): void {
+    const slicWatchConfig: SlicWatchConfig = this.serverless.service.custom?.slicWatch ?? {}
 
     const ajv = new Ajv({
       unicodeRegExp: false
@@ -57,19 +56,17 @@ class ServerlessPlugin {
     const slicWatchValidate = ajv.compile(slicWatchSchema)
     const slicWatchValid = slicWatchValidate(slicWatchConfig)
     if (!slicWatchValid) {
+      // eslint-disable-next-line @typescript-eslint/no-throw-literal
       throw new Serverless('SLIC Watch configuration is invalid: ' + ajv.errorsText(slicWatchValidate.errors))
     }
 
-    const alarmActions = []
+    const alarmActions: string[] = []
 
-    if (slicWatchConfig.ActionsEnabled === false) {
+    if (!(slicWatchConfig.enabled ?? true)) {
       return
     }
 
-    if (slicWatchConfig.topicArn) {
-      // @ts-ignore
-      alarmActions.push(slicWatchConfig.topicArn)
-    }
+    slicWatchConfig.topicArn != null && alarmActions.push(slicWatchConfig.topicArn)
 
     // Validate and fail fast on config validation errors since this is a warning in Serverless Framework 2.x
     const context = { alarmActions }
@@ -77,26 +74,19 @@ class ServerlessPlugin {
     const config = _.merge(defaultConfig, slicWatchConfig)
     const awsProvider = this.serverless.getProvider('aws')
 
-    const cfTemplate = CloudFormationTemplate(
-      this.serverless.service.provider.compiledCloudFormationTemplate,
-      this.serverless.service.resources ? this.serverless.service.resources.Resources : {}
-    )
-
     const functionAlarmConfigs = {}
     const functionDashboardConfigs = {}
     for (const funcName of this.serverless.service.getAllFunctions()) {
-      const func = this.serverless.service.getFunction(funcName)
-      // @ts-ignore
+      const func = this.serverless.service.getFunction(funcName) as any // check why they don't return slicWatch
       const functionResName = awsProvider.naming.getLambdaLogicalId(funcName)
-      // @ts-ignore
-      const funcConfig = func.slicWatch || {}
-      functionAlarmConfigs[functionResName] = funcConfig.alarms || {}
+      const funcConfig = func.slicWatch ?? {}
+      functionAlarmConfigs[functionResName] = funcConfig.alarms ?? {}
       functionDashboardConfigs[functionResName] = funcConfig.dashboard
     }
-    this.dashboard = dashboard(config.dashboard, functionDashboardConfigs)
-    this.alarms = alarms(config.alarms, functionAlarmConfigs, context)
-    this.dashboard.addDashboard(cfTemplate)
-    this.alarms.addAlarms(cfTemplate)
+    const compiledTemplate = this.serverless.service.provider.compiledCloudFormationTemplate
+    const additionalResources = this.serverless.service.resources as ResourceType
+    addDashboard(config.dashboard, functionDashboardConfigs, compiledTemplate, additionalResources)
+    addAlarms(config.alarms, functionAlarmConfigs, context, compiledTemplate, additionalResources)
   }
 }
 
